Fix postgres config formatting
