+++ autopg/__tests__/test_logic.py
@@ -20,44 +20,6 @@ from autopg.constants import (
 )
 from autopg.logic import Configuration, PostgresConfig
 
-#
-# New tests
-#
-
-
-@pytest.mark.parametrize(
-    "kb_value,expected",
-    [
-        # GB cases
-        (1048576, "1GB"),  # 1 GB
-        (2097152, "2GB"),  # 2 GB
-        (5242880, "5GB"),  # 5 GB
-        # MB cases
-        (1024, "1MB"),  # 1 MB
-        (2048, "2MB"),  # 2 MB
-        (51200, "50MB"),  # 50 MB
-        # KB cases
-        (1, "1kB"),  # 1 KB
-        (512, "512kB"),  # 512 KB
-        (1000, "1000kB"),  # Not quite 1 MB, should stay as KB
-        # Edge cases
-        (0, "0kB"),
-    ],
-)
-def test_format_kb_value(kb_value: int, expected: str):
-    """
-    Test the format_kb_value function with various inputs to ensure it correctly
-    formats values in kilobytes to the most appropriate unit (GB, MB, or kB).
-    """
-    from autopg.logic import format_kb_value
-
-    assert format_kb_value(kb_value) == expected
-
-
-#
-# Original tests
-#
-
 
 def test_is_configured_nothing_set():
     config = PostgresConfig(Configuration())
